feat: route back to creator landing page
